requirements.txt
# FastAPI and Web Framework
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.8.2
pydantic[email]==2.8.2
python-multipart==0.0.9

# Database Connections
pyodbc>=5.1.0

# Configuration
python-dotenv==1.0.1

# Security and Crypto
cryptography==42.0.5
bcrypt==4.1.3

# Caching
cachetools==5.3.3

# Utilities
python-dateutil==2.9.0
requests==2.31.0

# System monitoring
psutil==6.0.0

# Testing Dependencies
pytest==8.1.1
pytest-cov==5.0.0
pytest-asyncio==0.23.6
httpx==0.27.0